[ADDED] rock spawning
