+++ Vigilance-Ice-Skating-Racing/Assets/New Material.mat
@@ -73,6 +73,6 @@ Material:
     - _UVSec: 0
     - _ZWrite: 1
     m_Colors:
-    - _Color: {r: 0.5449448, g: 0.6403969, b: 0.735849, a: 1}
+    - _Color: {r: 0.85098046, g: 0.9176471, b: 0.9450981, a: 1}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
   m_BuildTextureStacks: []
