+++ Vigilance-Ice-Skating-Racing/Assets/Resources/Mountian_Obstacles.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 202b226fd562e9d4a9c6c658eef17cfa
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
