+++ Vigilance-Ice-Skating-Racing/Assets/Resources/Mountian_Obstacles/Rock1_Moutain.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3033adc43549cb44daadc3495344d8c9
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
