+++ Vigilance-Ice-Skating-Racing/Assets/Resources/Mountian_Obstacles/Rock2_Moutain.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 25785dbcc81881441b57bc3425156666
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
