+++ Vigilance-Ice-Skating-Racing/Assets/Resources/Mountian_Obstacles/Rock3_Mountian.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e82d914354e59aa4fb2927390aca4081
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
