+++ Vigilance-Ice-Skating-Racing/Assets/Scripts/TestConnection.cs
@@ -12,8 +12,8 @@ public class TestConnection : MonoBehaviour
     // Start is called before the first frame update
 
     private void Update(){
-        var p0 = a1.GetComponent<SurfaceGen>().EndPoint;
-        var p1 = a2.GetComponent<SurfaceGen>().StartPoint;
+        var p0 = a1.GetComponent<TerrainGeneration>().EndPoint;
+        var p1 = a2.GetComponent<TerrainGeneration>().StartPoint;
         p1.position = p0.position;
     }
 }
