+++ Vigilance-Ice-Skating-Racing/UserSettings/EditorUserSettings.asset
@@ -18,16 +18,16 @@ EditorUserSettings:
       value: 22424703114646680e0b0227036c73150012141e293b3239232e5326ece92021
       flags: 0
     RecentlyUsedScenePath-4:
-      value: 22424703114646680e0b0227036c72151802563f22213229
-      flags: 0
-    RecentlyUsedScenePath-5:
       value: 22424703114646680e0b0227036c72111f19563f22213229
       flags: 0
-    RecentlyUsedScenePath-6:
+    RecentlyUsedScenePath-5:
       value: 22424703114646680e0b0227036c7a1e121b1d393f1a333e232c0f7df7ee3d2cfb
       flags: 0
-    RecentlyUsedScenePath-7:
+    RecentlyUsedScenePath-6:
       value: 22424703114646680e0b0227036c6f1c170e1d38182d35241e2a183de7ae2136ebf32f
+      flags: 0
+    RecentlyUsedScenePath-7:
+      value: 22424703114646680e0b0227036c72151802563f22213229
       flags: 0
     RecentlyUsedScenePath-8:
       value: 22424703114646680e0b0227036c6b1505032c2f3e3a27392367083debf42d
